+++ processor/src/test/resources/expected/com/example/arez/lifecycle/React4j_ActionOnLifecycleComponent.java
@@ -1,0 +1,82 @@
+package com.example.arez.lifecycle;
+
+import arez.annotations.ArezComponent;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.ComponentConstructorFunction;
+import react4j.NativeAdapterComponent;
+import react4j.ReactConfig;
+
+@ArezComponent(
+    name = "ActionOnLifecycleComponent"
+)
+@Generated("react4j.processor.ReactProcessor")
+abstract class React4j_ActionOnLifecycleComponent extends ActionOnLifecycleComponent {
+  static final ComponentConstructorFunction TYPE = getConstructorFunction();
+
+  @Nonnull
+  private static ComponentConstructorFunction getConstructorFunction() {
+    final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "ActionOnLifecycleComponent" );
+    }
+    return componentConstructor;
+  }
+
+  @JsType(
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
+  )
+  interface Lifecycle {
+    void componentDidMount();
+
+    void componentDidUpdate(@Nonnull JsPropertyMap<Object> arg0,
+        @Nonnull JsPropertyMap<Object> arg1);
+
+    void componentWillUnmount();
+
+    boolean shouldComponentUpdate(@Nonnull JsPropertyMap<Object> arg0,
+        @Nonnull JsPropertyMap<Object> arg1);
+  }
+
+  private static final class NativeReactComponent extends NativeAdapterComponent<ActionOnLifecycleComponent> implements Lifecycle {
+    @JsConstructor
+    NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+    }
+
+    @Override
+    protected ActionOnLifecycleComponent createComponent() {
+      return new Arez_React4j_ActionOnLifecycleComponent();
+    }
+
+    @Override
+    public void componentDidMount() {
+      performComponentDidMount();
+    }
+
+    @Override
+    public void componentDidUpdate(@Nonnull final JsPropertyMap<Object> arg0,
+        @Nonnull final JsPropertyMap<Object> arg1) {
+      performComponentDidUpdate(arg0,arg1);
+    }
+
+    @Override
+    public void componentWillUnmount() {
+      performComponentWillUnmount();
+    }
+
+    @Override
+    public boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> arg0,
+        @Nonnull final JsPropertyMap<Object> arg1) {
+      return performShouldComponentUpdate(arg0,arg1);
+    }
+  }
+}